Move consts from webview to extension utils
