+++ packages/vscode-extension/src/webview/hooks/useResizableProps.tsx
@@ -1,6 +1,6 @@
 import { ResizeCallback } from "re-resizable";
 import { CSSProperties, useCallback, useEffect, useRef, useState } from "react";
-import { DeviceProperties } from "../utilities/consts";
+import { DeviceProperties } from "../../utilities/consts";
 import { ZoomLevelType } from "../../common/Project";
 import { DEVICE_DEFAULT_SCALE } from "../components/ZoomControls";
